Move shared extension logic into Extensions/common
Extracts state.js, utils.js, buttons.js, bar.js, config.js, events.js
(and their tests) out of Extensions/combined/src into a new
Extensions/common directory, so the browser-extension-specific code in
Extensions/combined stays limited to genuinely extension-only concerns
(chrome storage, Patreon OAuth, premium analytics UI, manifests). This
sets up Extensions/common as a single source of truth that the
upcoming userscript build will also pull from, instead of duplicating
this logic by hand.

No behavior change: all import paths were updated to point at the new
location, existing tests pass unmodified, and the extension build
output is unaffected.
