+++ Extensions/common/bar.js
