+++ Extensions/common/config.js
