+++ Extensions/common/events.js
