+++ Extensions/common/utils.js
